refactored code in order to add provider
