+++ pubspec.yaml
@@ -30,7 +30,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-
+  provider: ^6.1.2
   get_storage: ^2.1.1
 
   # The following adds the Cupertino Icons font to your application.
